basic_face_detection/requirements.txt
-r ../requirements.txt

dlib
face-recognition
numpy
Pillow
